Fix warning, still one unused variable but you didn't find a way to remove it

+++ exemple/main.c
@@ -8,15 +8,15 @@ static void* func(int *nbb)
 {
 	int nb = *nbb;
 	int i = 1;
-	printf("func%d: started (%d)\n", nb, thread_self());
+	printf("func%d: started (%d)\n", nb, (int)thread_self());
 	while(i < 7)
 	{
-		printf("while %d\n", thread_self());
+		printf("while %d\n", (int)thread_self());
 		thread_yield();
 		/*sleep(1);*/
 		++i;
 	}
-	printf("func%d: returning (%d)\n", nb, thread_self());
+	printf("func%d: returning (%d)\n", nb, (int)thread_self());
 	*nbb *= 3;
 	return (void*)nbb;
 }
@@ -36,7 +36,7 @@ int main(int argc, char *argv[])
 	int* value;
 	for(i = 0; i < SIZE; ++i)
 	{
-		printf("Joining %d i'm %d\n", thre[i], thread_self());
+		printf("Joining %d i'm %d\n", (int)thre[i], (int)thread_self());
 		thread_join(thre[i], (void**)&value);
 		printf("Result for %d : %d\n", i, *value);
 		free(value);
